pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "src"
version = "0.0.1"
description = "A short description of the project."
authors = [
  { name = "Mohamed Shakshak" },
]

readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    
]
dependencies = [
    "hydra-core>=1.3.2",
    "ipython",
    "jupyterlab",
    "kaggle>=2.0.1",
    "kagglehub>=1.0.0",
    "loguru",
    "matplotlib",
    "mkdocs",
    "notebook",
    "numpy",
    "omegaconf>=2.3.0",
    "pandas",
    "pip",
    "pytest",
    "python-dotenv",
    "ruff",
    "scikit-learn",
    "tqdm",
    "typer",
]
requires-python = "~=3.11.0"


[tool.ruff]
line-length = 99
src = ["src"]
include = ["pyproject.toml", "src/**/*.py"]

[tool.ruff.lint]
extend-select = ["I"]  # Add import sorting

[tool.ruff.lint.isort]
known-first-party = ["src"]
force-sort-within-sections = true

